dev: add: use aws sso for terraform-aws operations

+++ src/express/commands/destroy.js
@@ -5,14 +5,15 @@ const shell = require("shelljs");
 export async function terraformDestroy() {
     console.log("📍Executing terraform destroy...")
     require('dotenv').config({path: `${process.cwd()}/.env`})
-    shell.exec(`terraform destroy -auto-approve`, {
+    shell.exec(`terraform dumb -auto-approve`, {
         env: {
             ...process.env,
         }
     });
-
-    let devnetId = getDevnetId()
-    console.log(`📍Deleting ./deployments/devnet-${devnetId} folder...`)
-    shell.exec(`cd ../../ && rm -rf ./deployments/devnet-${devnetId}`)
-    console.log(`📍devnet-${devnetId} successfully destroyed!`)
+    if (shell.error() == null) {
+        let devnetId = getDevnetId()
+        console.log(`📍Deleting ./deployments/devnet-${devnetId} folder...`)
+        shell.exec(`cd ../../ && rm -rf ./deployments/devnet-${devnetId}`)
+        console.log(`📍devnet-${devnetId} successfully destroyed!`)
+    }
 }
